完善 Makfile

+++ Makefile
@@ -2,18 +2,19 @@
 # Use of this source code is governed by a BSD-style
 # license that can be found in the LICENSE file.
 
+.PHONY: default
 default: clean
 	hugo && mv public public-new && cd public-new && go run ../server.go
 
+.PHONY: debug
 debug:
 	hugo server --buildDrafts --watch
 
-caddy:
-	docker run --name chai2010.cn -d -p 80:80 -p 443:443 --restart=always -v /root/.caddy:/root/.caddy -v /root/.ssh:/root/.ssh chai2010/chai2010.cn
-
+.PHONY: pub
 pub:
 	# copy ./Caddyfile to ssh@/root/Caddyfile
 	# copy ./Makefile to ssh@/root/Makefile
+	-@docker kill chai2010.cn
 	docker run -d --restart=always --name chai2010.cn \
 		-v /root/Caddyfile:/etc/Caddyfile \
 		-v /root/.caddy:/root/.caddy \
@@ -21,15 +22,18 @@ pub:
 		-p 80:80 -p 443:443 \
 		abiosoft/caddy
 
+.PHONY: dev
 dev: clean
 	# run `hugo` command to generate ./public at first
-	docker run --rm --name chai2010.cn.dev \
+	-@docker kill chai2010.cn
+	hugo && docker run --rm --name chai2010.cn.dev \
 		-v `pwd`/Caddyfile.local:/etc/Caddyfile \
 		-v `pwd`/.caddy:/root/.caddy \
 		-v `pwd`:/srv/chai2010.cn \
 		-p 2015:2015 \
 		abiosoft/caddy
 
+.PHONY: clean
 clean:
 	-rm -rf public
 	-rm -rf public-new
